+++ src/stringify.ts
@@ -1,73 +1,37 @@
-import { spaces, entirelyValidIdentifier } from './shared';
+import { spaces } from './shared';
 import { StringifierOptions } from './interfaces';
 
 export function stringify(value: any, options?: StringifierOptions) {
-	const quote = (options && options.singleQuotes) ? "'" : '"';
 	const indentString = (options && options.spaces) ? spaces(options.spaces) : '\t';
-
-	return stringifyValue(value, quote, '\n', indentString, true);
-}
-
-// https://github.com/json5/json5/blob/65bcc556eb629984b33bb2163cbc10fba4597300/src/stringify.js#L110
-const escapeable: Record<string, string> = {
-	"'": "'",
-	'"': '"',
-	'\\': '\\',
-	'\b': 'b',
-	'\f': 'f',
-	'\n': 'n',
-	'\r': 'r',
-	'\t': 't',
-	'\v': 'v',
-	'\0': '0',
-	'\u2028': 'u2028',
-	'\u2029': 'u2029',
-};
-const escapeableRegex = /['"\\\b\f\n\r\t\v\0\u2028\u2029]/g;
-
-export function stringifyString(str: string, quote: string) {
-	const otherQuote = quote === '"' ? "'" : '"';
-	return quote + str.replace(escapeableRegex, char =>
-		char === otherQuote ? char : '\\' + escapeable[char]
-	) + quote;
+	return stringifyValue(value, '\n', indentString, true);
 }
 
 export function stringifyProperty(
 	key: string,
 	value: any,
-	quote: string,
 	indentation: string,
 	indentString: string,
 	newlines: boolean
 ): string {
 	return (
-		(entirelyValidIdentifier.test(key) ? key : stringifyString(key, quote)) +
-		': ' +
-		stringifyValue(value, quote, indentation, indentString, newlines)
+		JSON.stringify(key) + ': ' + stringifyValue(value, indentation, indentString, newlines)
 	);
 }
 
 export function stringifyValue(
 	value: any,
-	quote: string,
 	indentation: string,
 	indentString: string,
 	newlines: boolean
 ): string {
 	const type = typeof value;
 
-	if (type === 'string') {
-		return stringifyString(value, quote);
-	}
-
-	if (type === 'number' || type === 'boolean' || value === null)
-		return String(value);
-
-	if (Array.isArray(value)) {
+	if (type === 'boolean' || type === 'number' || type === 'string' || type === null) {
+		return JSON.stringify(value);
+	} else if (Array.isArray(value)) {
 		const elements = value.map(element =>
 			stringifyValue(
 				element,
-				quote,
 				indentation + indentString,
 				indentString,
 				true
@@ -83,15 +47,12 @@ export function stringifyValue(
 		}
 
 		return `[ ${elements.join(', ')} ]`;
-	}
-
-	if (type === 'object') {
+	} else if (type === 'object') {
 		const keys = Object.keys(value);
-		const properties = keys.map(key =>
+		const properties = keys.map(key => 
 			stringifyProperty(
 				key,
 				value[key],
-				quote,
 				indentation + indentString,
 				indentString,
 				newlines